refactor(qwen35): split config.rs into config/{model,tp,tokenizer}.rs
config.rs was 683 lines mixing three unrelated owners: model architecture
(Config35/LayerType/Rope/Text/Raw + constants + impl), tensor-parallel
(TensorParallelConfig), and tokenizer (AddedTokenConfig/Tokenizer*/effective_vocab).
Split into a flat config/ submodule tree: config.rs declares mod model/tokenizer/tp
and re-exports (pub(crate) use *), probe_config_json stays at the root, tests stay
at the root. Behavior-preserving: no logic/loader changes.

Validated on A100: cargo check + cargo fmt OK, cargo test --lib 95 passed /
0 failed (7 TP tests ignored: need 2 CUDA devices + Qwen3.5 weights).
config.rs 683 -> 290; +265 model / +70 tp / +89 tokenizer.
